+++ swing/common-ui/src/test/java/is/codion/swing/common/ui/component/text/TextFieldPanelTest.java
@@ -22,7 +22,7 @@ import is.codion.common.state.State;
 
 import org.junit.jupiter.api.Test;
 
-import static is.codion.swing.common.ui.Utilities.linkToEnabledState;
+import static is.codion.swing.common.ui.Utilities.enableComponents;
 import static org.junit.jupiter.api.Assertions.*;
 
 public class TextFieldPanelTest {
@@ -60,7 +60,7 @@ public class TextFieldPanelTest {
 		State enabledState = State.state();
 		TextFieldPanel inputPanel = TextFieldPanel.builder()
 						.build();
-		linkToEnabledState(enabledState, inputPanel);
+		enableComponents(enabledState, inputPanel);
 		assertFalse(inputPanel.textField().isEnabled());
 		assertFalse(inputPanel.button().isEnabled());
 		enabledState.set(true);