+++ changelog.md
@@ -17,6 +17,7 @@ Codion Change Log
 - TextPaneBuilder added.
 - FilterTable.ScrollToAdded.scrollToAddedRow() no longer calls moveLeadSelectionIndex(), caused a selection event.
 - Control.Builder.name() renamed caption().
+- Utilities.linkToEnabledState() renamed enableActions() and enableComponents().
 ### is.codion.swing.framework.ui
 - EntityApplicationPanel.Builder.applicationModelFactory() and applicationPanelFactory() renamed applicationModel() and applicationPanel().
 - EntityApplicationPanel.createLogControls() now protected final.